+++ aiobungie/internal/factory.py
@@ -23,7 +23,7 @@
 
 from __future__ import annotations
 
-__all__ = ("Factory",)
+__all__: tuple[str, ...] = ("Factory",)
 
 import collections.abc as collections
 import typing
@@ -59,15 +59,13 @@ if typing.TYPE_CHECKING:
 class Factory(interfaces.FactoryInterface):
     """The base deserialization factory class for all aiobungie data classes.
 
-    This factory is used to deserialize JSON responses from the REST client and turning them
+    Highly inspired hikari entity factory used to deserialize JSON responses from the REST client and turning them
     into a Python data classes object.
 
     This is only provided by `aiobungie.Client` base client.
     """
 
-    # This is kinda inspired by hikari's entity factory :p.
-
-    __slots__: collections.Sequence[str] = ("_net",)
+    __slots__ = ("_net",)
 
     def __init__(self, net: traits.Netrunner) -> None:
         self._net = net
@@ -156,6 +154,8 @@ class Factory(interfaces.FactoryInterface):
             icon=assets.Image(user_info.get("iconPath", assets.Image.partial())),
             types=memberships,
         )
+
+    # TODO: Rework this horibble function.
 
     # Deserialize a list of `destinyUserInfo`
     def deserialize_destiny_members(
@@ -1477,6 +1477,9 @@ class Factory(interfaces.FactoryInterface):
     def _deserialize_activity_values(
         self, payload: typedefs.JsonObject, /
     ) -> activity.ActivityValues:
+        team: typing.Optional[int] = None
+        if raw_team := payload.get("team"):
+            team = raw_team["basic"]["value"]
         return activity.ActivityValues(
             assists=payload["assists"]["basic"]["value"],
             deaths=payload["deaths"]["basic"]["value"],
@@ -1488,7 +1491,7 @@ class Factory(interfaces.FactoryInterface):
             kd_assists=payload["killsDeathsAssists"]["basic"]["value"],
             score=payload["score"]["basic"]["value"],
             duration=payload["activityDurationSeconds"]["basic"]["displayValue"],
-            team=payload["team"]["basic"]["value"],
+            team=team,
             completion_reason=payload["completionReason"]["basic"]["displayValue"],
             fireteam_id=payload["fireteamId"]["basic"]["value"],
             start_seconds=payload["startSeconds"]["basic"]["value"],
@@ -1498,7 +1501,9 @@ class Factory(interfaces.FactoryInterface):
         )
 
     def deserialize_activity(
-        self, payload: typedefs.JsonObject, /
+        self,
+        payload: typedefs.JsonObject,
+        /,
     ) -> activity.Activity:
         period = time.clean_date(payload["period"])
         details = payload["activityDetails"]
@@ -1508,7 +1513,11 @@ class Factory(interfaces.FactoryInterface):
         modes = [enums.GameMode(int(mode_)) for mode_ in details["modes"]]
         is_private = details["isPrivate"]
         membership_type = enums.MembershipType(int(details["membershipType"]))
+
+        # Since we're using the same fields for post activity method
+        # this check is required since post activity doesn't values values
         values = self._deserialize_activity_values(payload["values"])
+
         return activity.Activity(
             net=self._net,
             hash=ref_id,
@@ -1528,10 +1537,102 @@ class Factory(interfaces.FactoryInterface):
             self.deserialize_activity(activity_) for activity_ in payload["activities"]
         ]
 
+    def _deserialize_extended_weapon_values(
+        self, payload: typedefs.JsonObject
+    ) -> activity.ExtendedWeaponValues:
+        return activity.ExtendedWeaponValues(
+            reference_id=int(payload["referenceId"]),
+            kills=payload["values"]["uniqueWeaponKills"]["basic"]["value"],
+            precision_kills=payload["values"]["uniqueWeaponPrecisionKills"]["basic"][
+                "value"
+            ],
+            precision_kills_percentage=(
+                payload["values"]["uniqueWeaponKillsPrecisionKills"]["basic"]["value"],
+                payload["values"]["uniqueWeaponKillsPrecisionKills"]["basic"][
+                    "displayValue"
+                ],
+            ),
+        )
+
+    def _deserialize_extended_values(
+        self, payload: typedefs.JsonObject
+    ) -> activity.ExtendedValues:
+        weapons: typing.Optional[
+            collections.Collection[activity.ExtendedWeaponValues]
+        ] = None
+
+        if raw_weapons := payload.get("weapons"):
+            weapons = [
+                self._deserialize_extended_weapon_values(value) for value in raw_weapons
+            ]
+
+        return activity.ExtendedValues(
+            precision_kills=payload["values"]["precisionKills"]["basic"]["value"],
+            grenade_kills=payload["values"]["weaponKillsGrenade"]["basic"]["value"],
+            melee_kills=payload["values"]["weaponKillsMelee"]["basic"]["value"],
+            super_kills=payload["values"]["weaponKillsSuper"]["basic"]["value"],
+            ability_kills=payload["values"]["weaponKillsAbility"]["basic"]["value"],
+            weapons=weapons,
+        )
+
+    def _deserialize_post_activity_player(
+        self, payload: typedefs.JsonObject
+    ) -> activity.PostActivityPlayer:
+        return activity.PostActivityPlayer(
+            standing=int(payload["standing"]),
+            score=int(payload["score"]["basic"]["value"]),
+            character_id=payload["characterId"],
+            destiny_user=self.deserialize_destiny_user(payload["player"]),
+            character_class=payload["player"]["characterClass"],
+            character_level=int(payload["player"]["characterLevel"]),
+            race_hash=int(payload["player"]["raceHash"]),
+            gender_hash=int(payload["player"]["genderHash"]),
+            light_level=int(payload["player"]["lightLevel"]),
+            emblem_hash=int(payload["player"]["emblemHash"]),
+            class_hash=payload["player"]["classHash"],
+            values=self._deserialize_activity_values(payload["values"]),
+            extended_values=self._deserialize_extended_values(payload["extended"]),
+        )
+
+    def _deserialize_post_activity_team(
+        self, payload: typedefs.JsonObject
+    ) -> activity.PostActivityTeam:
+        return activity.PostActivityTeam(
+            id=payload["teamId"],
+            is_defeated=bool(payload["standing"]["basic"]["value"]),
+            score=int(payload["score"]["basic"]["value"]),
+            name=payload["teamName"],
+        )
+
     def deserialize_post_activity(
         self, payload: typedefs.JsonObject
     ) -> activity.PostActivity:
-        ...
+        period = time.clean_date(payload["period"])
+        details = payload["activityDetails"]
+        ref_id = int(details["referenceId"])
+        instance_id = int(details["instanceId"])
+        mode = enums.GameMode(details["mode"])
+        modes = [enums.GameMode(int(mode_)) for mode_ in details["modes"]]
+        is_private = details["isPrivate"]
+        membership_type = enums.MembershipType(int(details["membershipType"]))
+        return activity.PostActivity(
+            net=self._net,
+            hash=ref_id,
+            membership_type=membership_type,
+            instance_id=instance_id,
+            mode=mode,
+            modes=modes,
+            is_private=is_private,
+            occurred_at=period,
+            starting_phase=int(payload["startingPhaseIndex"]),
+            players=[
+                self._deserialize_post_activity_player(player)
+                for player in payload["entries"]
+            ],
+            teams=[
+                self._deserialize_post_activity_team(team) for team in payload["teams"]
+            ],
+        )
 
     def deserialize_linked_profiles(
         self, payload: typedefs.JsonObject